+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 
 ### Features 
 - Changed Referenced Var route_table_id = module.cato_deployment.lan_route_table_id to route_table_id = module.cato_deployment.lan_subnet_route_table_id
+- Updated Ref from AZID to AZ